implement all api routes

+++ api/serializers.py
@@ -17,7 +17,7 @@ class SeasonSerializer(serializers.ModelSerializer):
 
 
 class AnimeSerializer(serializers.ModelSerializer):
-    seasons = SeasonSerializer(many=True)
+    seasons = SeasonSerializer(many=True, required=False, default=[])
 
     class Meta:
         model = Anime
@@ -40,23 +40,23 @@ class AnimeSerializer(serializers.ModelSerializer):
         return anime
 
     def update(self, instance, validated_data):
-        seasons_data = validated_data.pop("seasons", [])
+        seasons_data = validated_data.pop("seasons", None)
 
         for attr, value in validated_data.items():
             setattr(instance, attr, value)
         instance.save()
 
-        instance.seasons.all().delete()
-        for season_data in seasons_data:
-            Season.objects.create(anime=instance, **season_data)
+        # only replace seasons when the key was actually sent
+        if seasons_data is not None:
+            instance.seasons.all().delete()
+            for season_data in seasons_data:
+                Season.objects.create(anime=instance, **season_data)
 
         return instance
 
 
 class CategorySerializer(serializers.ModelSerializer):
-    animes = AnimeSerializer(many=True)
-
     class Meta:
         model = Category
-        fields = ("id", "name", "order", "animes")
-        read_only_fields = "id"
+        fields = ("id", "name", "order")
+        read_only_fields = ("id",)
